move riscv canonicalization patterns to own folder
