+++ xdsl/transforms/canonicalization_patterns/riscv.py
